Refactored integration test namespace and class names
